// src/payload-types.ts
export interface Media {
  id: number;
  alt?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
// ...
export interface Project {
  id: number;
  slug: string;
  url?: string | null;
  category: number | Category;
  name: string;
  description: string;
  featuredImage?: (number | null) | Media;
  images?:
    | {
        image?: (number | null) | Media;
        id?: string | null;
      }[]
    | null;
  features?:
    | {
        feature: string;
        id?: string | null;
      }[]
    | null;
  duration?: string | null;
  role?: string | null;
  techStacks?:
    | {
        stack?: string | null;
        id?: string | null;
      }[]
    | null;
  results?:
    | {
        result: string;
        id?: string | null;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "categories".
 */
export interface Category {
  id: number;
  name: string;
  updatedAt: string;
  createdAt: string;
}
// ...
export interface Home {
  id: number;
  heroBanner: {
    title: string;
    description?: string | null;
    image?: (number | null) | Media;
  };
  services: {
    titleService: string;
    descService?: string | null;
    listService?:
      | {
          name: string;
          desc?: string | null;
          icon: number | Media;
          id?: string | null;
        }[]
      | null;
  };
  projects: {
    title: string;
    desc?: string | null;
    projects?:
      | {
          project?: (number | null) | Project;
          id?: string | null;
        }[]
      | null;
  };
  canIHelpYou: {
    title: string;
    desc?: string | null;
    utilities?:
      | {
          name: string;
          desc?: string | null;
          icon: number | Media;
          id?: string | null;
        }[]
      | null;
  };
  updatedAt?: string | null;
  createdAt?: string | null;
}